refactor: 연쇄 체인 패턴으로 정책을 추가하도록 리펙토링함
