+++ src/main/java/nextstep/subway/domain/policy/BasicFarePolicy.java
@@ -1,0 +1,7 @@
+package nextstep.subway.domain.policy;
+
+import nextstep.subway.domain.Path;
+
+public interface BasicFarePolicy {
+    int calculate(int age, int fare, int distance, Path path);
+}
