+++ src/main/java/nextstep/subway/domain/policy/distance/DistanceFarePolicy.java
@@ -1,0 +1,12 @@
+package nextstep.subway.domain.policy.distance;
+
+import nextstep.subway.domain.Path;
+import nextstep.subway.domain.policy.BasicFarePolicy;
+
+public class DistanceFarePolicy implements BasicFarePolicy {
+    @Override
+    public int calculate(int age, int fare, int distance, Path path) {
+        DistanceType policy = DistanceType.createFarePolicy(distance);
+        return policy.calculate(distance);
+    }
+}
